Fixes for make dist
cogl-handle.h was being listed in the sources but it has been removed
so the dist would not work.

Also the micro-perf directory under tests was not listed in
DIST_SUBDIRS so it would be skipped.

This patch was written by Rico Tzschichholz.

Reviewed-by: Neil Roberts <neil@linux.intel.com>

(cherry picked from commit fb96716aaffd72a00461ce4dd8496904de0bb1ff)